Don't overwrite local .gitconfig

+++ InstallApplications.ps1
@@ -91,7 +91,10 @@ Function InstallGit {
         '/NoGitLfs' # Disable Git LFS installation.
     )
 
-    Copy-Item -Path $PSScriptRoot\config\git\gitconfig.local -Destination $HOME\.gitconfig.local | Out-null
+    if(Test-Path $PSScriptRoot\config\git\gitconfig.local -ne $null)
+    {
+        Copy-Item -Path $PSScriptRoot\config\git\gitconfig.local -Destination $HOME\.gitconfig.local | Out-null
+    }
     New-Item -ItemType HardLink -Force -Path $HOME -Name .gitignore -Value $PSScriptRoot\config\git\gitignore | Out-null
     New-Item -ItemType HardLink -Force -Path $HOME -Name .gitattributes -Value $PSScriptRoot\config\git\gitattributes | Out-null
     New-Item -ItemType HardLink -Force -Path $HOME -Name .gitconfig -Value $PSScriptRoot\config\git\gitconfig | Out-null
